ImageGen.py in Scorer used, one image function moved to MedusaMongo
