+++ Medusa/ImageGenerator.py
@@ -1,8 +1,9 @@
-import numpy
-from PIL import Image
+from numpy import random
+import io
+from PIL import Image, ImageEnhance
 
 def create_image(width = 1920, height = 1080 , name = 'random.png'):
-    imarray = numpy.random.rand(width,height,3) * 255
+    imarray = random.rand(width,height,3) * 255
     im = Image.fromarray(imarray.astype('uint8')).convert('RGBA')
     im.save(name)
 
@@ -15,7 +16,7 @@ def create_image_with_color_prop(width = 64, height = 64, prop_red = 33, prop_gr
 
 def create_bytearray_with_color_prop(width = 64, height = 64, prop_red = 33, prop_green = 34, prop_blue = 33):
     #Initalize array with zeros
-    imarray = numpy.random.rand(width * height, 3) * 0
+    imarray = random.rand(width * height, 3) * 0
     red_pixel = prop_red/100*width*height
     green_pixel = prop_green/100*width*height
     blue_pixel = prop_blue/100*width*height
@@ -34,8 +35,8 @@ def create_bytearray_with_color_prop(width = 64, height = 64, prop_red = 33, pro
                 #Wrong proportion --> white pixel
                 imarray[x*width+y] = [255,255,255]
     
-    numpy.random.shuffle(imarray)
-    immatrix = numpy.random.rand(height, width, 3) * 0
+    random.shuffle(imarray)
+    immatrix = random.rand(height, width, 3) * 0
     for x in range(height):
         for y in range(width):
             immatrix[x,y] = imarray[x*width+y]
@@ -53,3 +54,15 @@ def change_contrast_of_img(image, contrastFactor=1.0):
 def create_img_from_bytearray(bytearr, colorscheme='RGBA'):
     im=Image.fromarray(bytearr).convert(colorscheme)
     return im
+
+#From Scorer
+def create_bytearray(width=64,height=64):
+    imarray = random.rand(width,height,3)*255
+    casted = imarray.astype('uint8')
+    return casted
+
+def img_to_bytearray(Image):
+    imgByteArr = io.BytesIO()
+    Image.save(imgByteArr, format='PNG')
+    imgByteArr = imgByteArr.getvalue()
+    return imgByteArr
